+++ src/main/java/com/arangodb/ArangoDB.java
@@ -20,14 +20,25 @@
 
 package com.arangodb;
 
-import com.arangodb.entity.*;
+import com.arangodb.entity.ArangoDBEngine;
+import com.arangodb.entity.ArangoDBVersion;
+import com.arangodb.entity.LoadBalancingStrategy;
+import com.arangodb.entity.LogEntity;
+import com.arangodb.entity.LogLevelEntity;
+import com.arangodb.entity.Permissions;
+import com.arangodb.entity.ServerRole;
+import com.arangodb.entity.UserEntity;
 import com.arangodb.internal.ArangoContext;
 import com.arangodb.internal.ArangoDBImpl;
 import com.arangodb.internal.ArangoDefaults;
 import com.arangodb.internal.InternalArangoDBBuilder;
 import com.arangodb.internal.http.HttpCommunication;
 import com.arangodb.internal.http.HttpConnectionFactory;
-import com.arangodb.internal.net.*;
+import com.arangodb.internal.net.ConnectionFactory;
+import com.arangodb.internal.net.Host;
+import com.arangodb.internal.net.HostHandle;
+import com.arangodb.internal.net.HostHandler;
+import com.arangodb.internal.net.HostResolver;
 import com.arangodb.internal.util.ArangoDeserializerImpl;
 import com.arangodb.internal.util.ArangoSerializationFactory;
 import com.arangodb.internal.util.ArangoSerializerImpl;
@@ -42,7 +53,18 @@ import com.arangodb.util.ArangoCursorInitializer;
 import com.arangodb.util.ArangoDeserializer;
 import com.arangodb.util.ArangoSerialization;
 import com.arangodb.util.ArangoSerializer;
-import com.arangodb.velocypack.*;
+import com.arangodb.velocypack.VPack;
+import com.arangodb.velocypack.VPackAnnotationFieldFilter;
+import com.arangodb.velocypack.VPackAnnotationFieldNaming;
+import com.arangodb.velocypack.VPackDeserializer;
+import com.arangodb.velocypack.VPackInstanceCreator;
+import com.arangodb.velocypack.VPackJsonDeserializer;
+import com.arangodb.velocypack.VPackJsonSerializer;
+import com.arangodb.velocypack.VPackModule;
+import com.arangodb.velocypack.VPackParser;
+import com.arangodb.velocypack.VPackParserModule;
+import com.arangodb.velocypack.VPackSerializer;
+import com.arangodb.velocypack.ValueType;
 import com.arangodb.velocystream.Request;
 import com.arangodb.velocystream.Response;
 import org.apache.http.client.HttpRequestRetryHandler;
@@ -245,6 +267,19 @@ public interface ArangoDB extends ArangoSerializationAccessor {
          */
         public Builder connectionTtl(final Long connectionTtl) {
             setConnectionTtl(connectionTtl);
+            return this;
+        }
+
+        /**
+         * Set the keep-alive interval for VST connections. If set, every VST connection will perform a no-op request every
+         * {@code keepAliveInterval} seconds, to avoid to be closed due to inactivity by the server (or by the external
+         * environment, eg. firewall, intermediate routers, operating system).
+         *
+         * @param keepAliveInterval interval in seconds
+         * @return {@link ArangoDB.Builder}
+         */
+        public Builder keepAliveInterval(final Integer keepAliveInterval) {
+            setKeepAliveInterval(keepAliveInterval);
             return this;
         }
 
@@ -600,7 +635,7 @@ public interface ArangoDB extends ArangoSerializationAccessor {
             final int max = maxConnections != null ? Math.max(1, maxConnections) : protocolMaxConnections;
 
             final ConnectionFactory connectionFactory = (protocol == null || Protocol.VST == protocol)
-                    ? new VstConnectionFactorySync(host, timeout, connectionTtl, useSsl, sslContext)
+                    ? new VstConnectionFactorySync(host, timeout, connectionTtl, keepAliveInterval, useSsl, sslContext)
                     : new HttpConnectionFactory(timeout, user, password, useSsl, sslContext, hostnameVerifier, custom,
                     protocol, connectionTtl, httpCookieSpec, httpRequestRetryHandler);